# docs/api/swagger.yaml
basePath: /
definitions:
  github_com_zhian9_blogo-server_pkg_errors.Error:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      detail:
        description: 详细描述
        type: string
      id:
        description: 错误标识（前端友好）
        type: string
      status:
        description: 状态文本（如 "Not Found"）
        type: string
    type: object
...
  schema.FriendLink:
    properties:
      created_at:
        description: 创建时间
        type: string
      description:
        description: 站点描述
        type: string
      email:
        description: 联系邮箱（用于审核沟通）
        type: string
      id:
        description: 友链唯一ID
        type: string
      logo:
        description: Logo图片URL（可关联 Image 表）
        type: string
      name:
        description: 站点名称
        type: string
      sort:
        description: 排序值（越小越靠前）
        type: integer
      status:
        description: 状态：enabled / disabled
        type: string
      updated_at:
        description: 更新时间
        type: string
      url:
        description: 站点链接
        type: string
    type: object
  schema.FriendLinkForm:
    properties:
      description:
        description: 描述
        maxLength: 255
        type: string
      email:
        description: 联系邮箱
        maxLength: 128
        type: string
      logo:
        description: Logo图片URL
        maxLength: 512
        type: string
      name:
        description: 站点名称
        maxLength: 100
        type: string
      sort:
        description: 排序值
        minimum: 0
        type: integer
      status:
        description: 状态
        enum:
        - enabled
        - disabled
        type: string
      url:
        description: 站点链接
        type: string
    required:
    - name
    - status
    - url
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_zhian9_blogo-server_pkg_errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
host: localhost:8040
info:
  contact:
    email: lxyaa911@gmail.com
    name: 李星云 (lxy911)
    url: https://github.com/zhian9
  description: Modern SaaS Blog Platform — REST API for content management, user system,
    RBAC, analytics, and more.
  license:
    name: MIT
    url: https://github.com/zhian9/Blogo/blob/master/LICENSE
  termsOfService: https://github.com/zhian9/Blogo
  title: Blogo API
  version: v1.0.0
paths:
...
  /api/v1/friend-links:
    get:
      parameters:
      - default: 1
        description: pagination index
        in: query
        name: current
        type: integer
      - default: 10
        description: pagination size
        in: query
        name: pageSize
        type: integer
      - description: Friend link name (fuzzy)
        in: query
        name: name
        type: string
      - description: Status
        enum:
        - enabled
        - disabled
        in: query
        name: status
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.FriendLink'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Query friend link list
      tags:
      - FriendLinkAPI
    post:
      parameters:
      - description: Friend link form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.FriendLinkForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.FriendLink'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Create friend link
      tags:
      - FriendLinkAPI
  /api/v1/friend-links/{id}:
    delete:
      parameters:
      - description: Friend link ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Delete friend link
      tags:
      - FriendLinkAPI
    get:
      parameters:
      - description: Friend link ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.FriendLink'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Get friend link by ID
      tags:
      - FriendLinkAPI
    put:
      parameters:
      - description: Friend link ID
        in: path
        name: id
        required: true
        type: string
      - description: Friend link form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.FriendLinkForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Update friend link
      tags:
      - FriendLinkAPI
  /api/v1/friend-links/all:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.FriendLink'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Get all enabled friend links (public)
      tags:
      - FriendLinkAPI
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your JWT token>". Login via /api/v1/login to obtain
      a token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"